requirements.txt
pandas==2.1.4
streamlit==1.29.0
requests==2.31.0
openai==1.6.1
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.17.4
python-dotenv==1.0.0
numpy==1.26.2
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.4
chromadb==0.4.21